Moved stream controllers into flags controller so clients are updated
