+++ routes/index.js
@@ -2,16 +2,11 @@ const express = require("express");
 const router = express.Router();
 
 const flagsController = require("../controllers/flagsController");
-const streamController = require("../controllers/streamController");
 
-router.post(
-  "/flags",
-  flagsController.replaceFlags
-  // streamController.sendUpdate
-);
+router.post("/flags", flagsController.replaceFlags);
 
 router.get("/flags", flagsController.getFlags);
 
-router.get("/stream", streamController.handleNewConnection);
+router.get("/stream", flagsController.handleNewConnection);
 
 module.exports = router;
